authorization

+++ API/Controllers/WebsiteController.cs
@@ -1,4 +1,5 @@
-﻿using API.Dto;
+﻿using System.Security.Claims;
+using API.Dto;
 using API.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -20,6 +21,16 @@ public class WebsiteController : ControllerBase
     [HttpPost("create/{userId}")]
     public async Task<IActionResult> CreateWebsite(int userId, WebsiteCreateDto websiteCreateDto)
     {
+        
+        var tokenUserId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        // Check if the userId in the route matches the one in the token
+        if (userId != tokenUserId)
+        {
+            return Unauthorized("Unauthorized access. User ID mismatch.");
+        }
+        
+        
         try
         {
             var website = await _websiteService.CreateWebsiteAsync(userId, websiteCreateDto);
@@ -36,6 +47,15 @@ public class WebsiteController : ControllerBase
     [HttpGet("getWebsites/{userId}")]
     public async Task<IActionResult> GetWebsitesByUserId(int userId)
     {
+        
+        var tokenUserId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        // Check if the userId in the route matches the one in the token
+        if (userId != tokenUserId)
+        {
+            return Unauthorized("Unauthorized access. User ID mismatch.");
+        }
+        
         try
         {
             var websites = await _websiteService.GetWebsitesByUserIdAsync(userId);
@@ -55,6 +75,15 @@ public class WebsiteController : ControllerBase
     [HttpDelete("delete/{userId}/{websiteId}")]
     public async Task<IActionResult> DeleteWebsite(int userId, int websiteId)
     {
+        
+        var tokenUserId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        // Check if the userId in the route matches the one in the token
+        if (userId != tokenUserId)
+        {
+            return Unauthorized("Unauthorized access. User ID mismatch.");
+        }
+        
         try
         {
             await _websiteService.DeleteWebsiteAsync(userId, websiteId);
